PUBDEV-5742 Extra benchmarks for validation frame early stopping scenario.
