+++ h2o-automl/src/main/java/ai/h2o/automl/AutoML.java
@@ -25,6 +25,7 @@ import water.exceptions.H2OIllegalArgumentException;
 import water.fvec.Frame;
 import water.fvec.Vec;
 import water.nbhm.NonBlockingHashMap;
+import water.rapids.StratificationAssistant;
 import water.util.ArrayUtils;
 import water.util.Countdown;
 import water.util.IcedHashMapGeneric;
